Frontend: Allow recipient registry type to be set at build time
