+++ vue-app/.env.example
@@ -3,3 +3,4 @@ VUE_APP_BLOCK_EXPLORER=https://etherscan.io/tx/
 VUE_APP_IPFS_GATEWAY_URL=https://ipfs.io
 VUE_APP_CLRFUND_FACTORY_ADDRESS=0x3619DbE27d7c1e7E91aA738697Ae7Bc5FC3eACA5
 VUE_APP_USER_REGISTRY_TYPE=simple
+VUE_APP_RECIPIENT_REGISTRY_TYPE=simple
