+++ vue-app/src/api/projects.ts
@@ -1,8 +1,6 @@
-import { Contract, Event } from 'ethers'
-import { isAddress } from '@ethersproject/address'
+import { factory, recipientRegistryType } from './core'
 
-import { SimpleRecipientRegistry } from './abi'
-import { factory, provider, ipfsGatewayUrl } from './core'
+import SimpleRegistry from './recipient-registry-simple'
 
 export interface Project {
   id: string; // Address or another ID depending on registry implementation
@@ -13,76 +11,23 @@ export interface Project {
   isRemoved: boolean;
 }
 
-function decodeRecipientAdded(event: Event): Project {
-  const args = event.args as any
-  const metadata = JSON.parse(args._metadata)
-  return {
-    id: args._recipient,
-    name: metadata.name,
-    description: metadata.description,
-    imageUrl: `${ipfsGatewayUrl}/ipfs/${metadata.imageHash}`,
-    index: args._index.toNumber(),
-    isRemoved: false,
-  }
-}
-
 export async function getProjects(
   startBlock?: number,
   endBlock?: number,
 ): Promise<Project[]> {
   const registryAddress = await factory.recipientRegistry()
-  const registry = new Contract(registryAddress, SimpleRecipientRegistry, provider)
-  const recipientAddedFilter = registry.filters.RecipientAdded()
-  const recipientAddedEvents = await registry.queryFilter(recipientAddedFilter, 0, endBlock)
-  const recipientRemovedFilter = registry.filters.RecipientRemoved()
-  const recipientRemovedEvents = await registry.queryFilter(recipientRemovedFilter, 0)
-  const projects: Project[] = []
-  for (const event of recipientAddedEvents) {
-    let project
-    try {
-      project = decodeRecipientAdded(event)
-    } catch {
-      // Invalid metadata
-      continue
-    }
-    const removed = recipientRemovedEvents.find((event) => {
-      return (event.args as any)._recipient === project.id
-    })
-    if (removed) {
-      if (!startBlock || startBlock && removed.blockNumber <= startBlock) {
-        // Start block not specified
-        // or recipient had been removed before start block
-        continue
-      } else {
-        project.isRemoved = true
-      }
-    }
-    projects.push(project)
+  if (recipientRegistryType === 'simple') {
+    return await SimpleRegistry.getProjects(registryAddress, startBlock, endBlock)
+  } else {
+    throw new Error('invalid recipient registry type')
   }
-  return projects
 }
 
-export async function getProject(address: string): Promise<Project | null> {
-  if (!isAddress(address)) {
-    return null
-  }
+export async function getProject(id: string): Promise<Project | null> {
   const registryAddress = await factory.recipientRegistry()
-  const registry = new Contract(registryAddress, SimpleRecipientRegistry, provider)
-  const recipientAddedFilter = registry.filters.RecipientAdded(address)
-  const recipientAddedEvents = await registry.queryFilter(recipientAddedFilter, 0)
-  if (recipientAddedEvents.length !== 1) {
-    return null
+  if (recipientRegistryType === 'simple') {
+    return await SimpleRegistry.getProject(registryAddress, id)
+  } else {
+    throw new Error('invalid recipient registry type')
   }
-  let project
-  try {
-    project = decodeRecipientAdded(recipientAddedEvents[0])
-  } catch {
-    return null
-  }
-  const recipientRemovedFilter = registry.filters.RecipientRemoved(address)
-  const recipientRemovedEvents = await registry.queryFilter(recipientRemovedFilter, 0)
-  if (recipientRemovedEvents.length !== 0) {
-    project.isRemoved = true
-  }
-  return project
 }
